move all conversion and decoding to utils.convert_and_decode
