+++ cdm_reader_mapper/mdf_reader/utils/__init__.py
@@ -1,6 +1,3 @@
 """Common Data Model (CDM) reader utilities."""
 
 from __future__ import annotations
-
-from .converters import converters  # noqa
-from .decoders import decoders  # noqa
